Move ofTools into ROSCO_toolbox folder for consistency
