+++ ROSCO_toolbox/ofTools/fast_io/FAST_reader.py
@@ -4,7 +4,7 @@ import numpy as np
 from functools import reduce
 import operator
 
-from ofTools.fast_io.FAST_vars import FstModel
+from ROSCO_toolbox.ofTools.fast_io.FAST_vars import FstModel
 from ROSCO_toolbox.utilities import read_DISCON, load_from_txt
 from ROSCO_toolbox import turbine as ROSCO_turbine
 ROSCO = True
