+++ ROSCO_toolbox/ofTools/fast_io/FAST_writer.py
@@ -4,8 +4,8 @@ import yaml
 import numpy as np
 from functools import reduce
 
-from ofTools.fast_io.FAST_reader import InputReader_Common, InputReader_OpenFAST, InputReader_FAST7
-from ofTools.fast_io.FAST_vars import FstModel
+from ROSCO_toolbox.ofTools.fast_io.FAST_reader import InputReader_Common, InputReader_OpenFAST, InputReader_FAST7
+from ROSCO_toolbox.ofTools.fast_io.FAST_vars import FstModel
 
 from ROSCO_toolbox.utilities import write_rotor_performance, write_DISCON
 ROSCO = True
