+++ ROSCO_toolbox/ofTools/util/spectral.py
